forwarder/requirements.txt
boto3==1.35.60
functions-framework==3.*
google-auth==2.38.0
google-cloud-logging==3.11.4
requests==2.32.3
